templates/requirements/base.txt
<% if (dbMysql) { %>
# Database Adapter
MySQL-python
<% } else if (dbPostgres) { %>
# Database Adapter
psycopg2
<% } %>

# Django stuff
Django

# Templates
django_compressor

# Celery: Message queue
celery
django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Logging
logutils

# Mixins
django-braces
django-model-utils
bpython

# Admin
django-debug-toolbar

# utils
path.py
dj-database-url
